Add file Delete Function

+++ Project1-2/controllers/admin.js
@@ -1,4 +1,6 @@
 const Product = require('../models/product');
+const fileHelper = require('../util/file');
+
 const { validationResult } = require('express-validator/check');
 
 exports.getAddProduct = (req, res, next) => {
@@ -143,6 +145,7 @@ exports.postEditProduct = (req, res, next) => {
       product.description = updatedDesc;
       // 새로운 이미지가 있으면 없데이트
       if (image) {
+        fileHelper.deleteFile(product.imageUrl);
         product.imageUrl = image.path;
       }
       return product.save().then(result => {
@@ -179,8 +182,14 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  // product id와 userId가 모두 같아야 삭제 가능
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
+  Product.findById(prodId)
+    .then(product => {
+      if (!product) {
+        return next(new Error('Product not Found.'));
+      }
+      fileHelper.deleteFile(product.imageUrl);
+      return Product.deleteOne({ _id: prodId, userId: req.user._id });
+    })
     .then(result => {
       console.log(result);
       console.log('DESTROYED PRODUCT');
